Renamed services to patterns, controller to ui. Added sharable link
