+++ src/client/public/js/patterns.js
