+++ src/client/views/index.pug
@@ -70,6 +70,6 @@ html(lang='en')
             a(href="https://github.com/goatonabicycle/note-generation", target="_blank") Code here
             br                       
 
-            script(src="js/controller.js", type="module")
+            script(src="js/ui.js", type="module")
